requirements.txt
datasets==1.2.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
diskcache==5.2.1
distributed==2021.9.1
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.10.0
future==0.18.2
gast==0.3.3
GDAL==3.4.0
girder-client==3.1.5
girder-slicer-cli-web==1.2.1
gitdb==4.0.5
GitPython==3.1.9
Glymur==0.9.4
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.18
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
histomicstk==1.1.0
idna==2.10
imageio==2.9.0
img2vec-pytorch==0.2.5
imgaug==0.4.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-contrib-core==0.3.3
jupyter-core==4.6.3
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
large-image==1.8.2
large-image-source-bioformats==1.8.2
large-image-source-deepzoom==1.8.2
large-image-source-dummy==1.8.2
large-image-source-gdal==1.8.2
large-image-source-mapnik==1.8.2
large-image-source-nd2==1.8.2
large-image-source-ometiff==1.8.2
large-image-source-openjpeg==1.8.2
large-image-source-openslide==1.8.2
large-image-source-pil==1.8.2
large-image-source-test==1.8.2
large-image-source-tiff==1.8.2
lazy-object-proxy==1.4.3
libtiff==0.5.0
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
mapnik==4.0.0
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
memory-profiler==0.58.0
mistune==0.8.4
msgpack==1.0.2
multiprocess==0.70.11.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nbresuse==0.3.6
nd2reader==3.3.0
nest-asyncio==1.4.2
networkx==2.5
nimfa==1.4.0
notebook==6.1.4
nptyping==1.4.1
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opencv-python-headless==4.5.3.56
openslide-python==1.1.2
opt-einsum==3.3.0
packaging==20.4
palettable==3.3.0
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
partd==1.2.0
pathos==0.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
PIMS==0.5
pkg-resources==0.0.0
pkgconfig==1.5.5
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.2
Pygments==2.7.2
pylint==2.6.0
pynndescent==0.5.2
pyparsing==2.4.7
pyproj==3.2.1.1
pyrsistent==0.17.3
python-bioformats==4.0.5
python-dateutil==2.8.1
python-javabridge==4.0.3
pytz==2020.1
pyvips==2.1.15
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.5.0
sacremoses==0.0.43
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
Shapely==1.7.1
SimpleITK==2.1.1
six==1.15.0
slicerator==1.0.0
smmap==3.0.4
sortedcontainers==2.4.0
SQLAlchemy==1.4.25
tblib==1.7.0
tdqm==0.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tifftools==1.2.8
tokenizers==0.9.4
toml==0.10.1
toolz==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.0.4
tqdm==4.49.0
traitlets==5.0.5
transformers==4.2.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typish==1.9.1
umap-learn==0.5.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
xxhash==2.0.0
yattag==1.14.0
zict==2.0.0
zipp==3.4.0
